multi_agent_system/requirements.txt
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
aiofiles>=23.0.0
transformers>=4.21.0
torch>=2.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0